+++ NEWS.md
@@ -1,7 +1,7 @@
 # ecotraj 0.2.1
 * Update of function 'centerTrajectories' to add flexibility in trajectory centroid definition
-* New vignette to illustrate centering
-* Improvement of trajectoryProjection to return relative position within selected segments
+* New vignette to illustrate centering.
+* Improvement of trajectoryProjection() to return relative position within selected segments, also when orthogonal projection does not exist.
 
 # ecotraj 0.2.0
 * New function 'trajectorySmoothing' added